moved getLastTableUpdate to new trdm package with unit tests
